refactor(orlando): move o fallback de logo para módulo próprio
Review do React Doctor na PR #1347: exportar uma função não-componente de
um arquivo de componente impede o Fast Refresh de preservar estado — o
editor passa a fazer reload completo a cada alteração no arquivo.

O handler foi exportado no commit anterior justamente para poder ser
testado; agora mora em logoFallback.ts, o que mantém o teste e devolve o
Fast Refresh. Mesmo padrão de constants.ts/assetPath.ts na landing irmã.
